reworked log plots. Default to symbol.
The default marker is now a symbol.

Factored out the code to deal with log plots and plot ranges.

Allow type= as a shorthand, with the marker type determined by the
value.  The other shorthands are dealt with when the maker is created,
e.g. Points or Curve

Allow scalar input

Documented the changes