warn of irrecoverable error when X is not available
It appears that if pl_openpl_r fails and the device is X, a segfault is
unavoidable.  It seems internally some kind of memory corruption is
occurring, or perhaps not being cleaned up.

I suspect this is a bug in plotutils itself.  The writers expect us
to exit(1) rather than clean up like a library should.

The problem can occur even when X is available; sometimes there seems to
be some kind of X auth problem on many systems where it appears X is
temporarily not available (but it is).

This also results in failure at pl_openpl_r.

The result is segfault.  In an interactive session this is highly
annoying, because all work to that point will be lost.

This may finally spell the death of biggles for me.