plot key enhancements
PlotKey can now be created without a list of components, and
then it can be added to later with .add().

a new argument can be sent to FramedPlot(key=), sending a plotkey
instance.  When objects are added to the plot, they are also added
to the plot key if they have a label attribute